Improve Sphinx configuration
With this change, all required coniguration to use this extension is
done when the extension is set up. This means no configuration needs to
be done in the conf.py. If values that this extension needs to set are
set by the user in conf.py (or by other extensions), warnings are
issued.