Do a more sane URL comparison
Previously, the src or href attributes were compared directly as
strings. Since the version of Bokeh is encoded in the URL, this caused a
test failure when the installed version of Bokeh was bumped. The
compatibility of versions is handled in setup.cfg, so this change
ignores the version part of the URL if it has Bokeh in it.

In addition, this change adds a check that BokehJS is included
_somewhere_ in the file (since that's the whole point of this package).